Fix for bool mask output

+++ src/models/golden_muscat_model.py
@@ -96,7 +96,7 @@ class golden_muscat_model(pipeline_pytorch_model):
             with torch.no_grad():
                 for i in range(0, len(norm_data), batch_size):
                     prediction = self.model.model(norm_patches[i:i+batch_size], legend_patches[i:i+batch_size])
-                    prediction = torch.argmax(prediction, dim=1).cpu().numpy().astype(np.uint8)
+                    prediction = torch.argmax(prediction, dim=1).cpu().numpy()
                     
                     prediction_patches.append(prediction)
                     
@@ -143,7 +143,7 @@ class golden_muscat_model(pipeline_pytorch_model):
         cur_max = cur_max
 
         for k in predictions:
-            predictions[k] = (predictions[k] >= cur_max) & (cur_max > 0.5)
+            predictions[k] = ((predictions[k] >= cur_max) & (cur_max > 0.5)).astype(np.uint8)
 
         return predictions
